Unify next/prev item selection
