+++ app/javascript/mastodon/components/hotkeys/index.tsx
@@ -270,9 +270,15 @@ export const Hotkeys: React.FC<{
    * Allow the rendered `div` to be focused
    */
   focusable?: boolean;
-  children: React.ReactNode;
+  children:
+    | React.ReactNode
+    | ((ref: React.RefObject<HTMLDivElement>) => React.ReactNode);
 }> = ({ handlers, global, focusable = true, children }) => {
   const ref = useHotkeys<HTMLDivElement>(handlers);
+
+  if (typeof children === 'function') {
+    return children(ref);
+  }
 
   return (
     <div ref={global ? undefined : ref} tabIndex={focusable ? -1 : undefined}>
